front: unify DeployedWaypoint type
This type was repeated thrice. Move it alongside its
MovableOccupancyZone sibling.
